Add "Managed Runtime Systems" course
"Managed Runtime Systems" is essentially a spinoff of Mario's "Virtual Machines and Managed Runtimes, UCB CS294" removing some of the historic info in an effort to spend more time on technical discussions and examples in the classroom. 
Disclaimer: About half (if not more) of the material is inherited (with attribution) from Mario.